+++ src/imio/smartweb/core/contents/__init__.py
@@ -4,6 +4,8 @@ from .folder.content import IFolder, Folder  # NOQA
 from .pages.footer.content import IFooter, Footer  # NOQA
 from .pages.page.content import IPage, Page  # NOQA
 from .pages.procedure.content import IProcedure, Procedure  # NOQA
+from .rest.events.content import IEventsView, EventsView  # NOQA
+from .rest.news.content import INewsView, NewsView  # NOQA
 from .sections.base import ISection, Section  # NOQA
 from .sections.contact.content import ISectionContact, SectionContact  # NOQA
 from .sections.files.content import ISectionFiles, SectionFiles  # NOQA