+++ src/imio/smartweb/core/contents/rest/news/content.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+
+from plone.dexterity.content import Container
+from plone.supermodel import model
+from zope.interface import implementer
+
+
+class INewsView(model.Schema):
+    """ """
+
+
+@implementer(INewsView)
+class NewsView(Container):
+    """NewsView class"""