+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 - Change navigation markup (quickaccess, close / prev buttons, ...)
   [laulaz]
+- Add two new types : imio.smartweb.EventsView, imio.smartweb.NewsView
+  [boulch]
 
 - Fix schedule display in Contact section (days delta, format & translations)
   [laulaz]